requirements.txt
acres==0.2.0
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.1
markdownify==0.14.1
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.2
numpy==2.0.2
openai==1.59.7
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
